a bug of merge

+++ src/config.py
@@ -1,5 +1,5 @@
 import torch
-DTYPE = torch.float32
+DTYPE = torch.float16
 # FILE_PATH = "stablediffusionapi/anything-v5"
 FILE_PATH = "CompVis/stable-diffusion-v1-4"
 DEVICE = torch.device("cuda")
